server/requirements.txt
flask==3.0.2
flask-cors==4.0.0
python-dotenv==1.0.1
gTTS==2.5.1
numpy==1.26.4
pillow==10.2.0
requests==2.31.0
